pyproject.toml
[project]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "python-chess (>=1.999,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "coverage (>=7.11.3,<8.0.0)",
    "invoke (>=2.2.1,<3.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pylint (>=4.0.3,<5.0.0)",
    "autopep8 (>=2.3.2,<3.0.0)"
]
